+++ integration-test/src/test/java/fi/vincit/multiusertest/ComponentInheritanceTest.java
@@ -21,7 +21,7 @@ public class ComponentInheritanceTest extends TestBaseClass {
 
     @Test
     public void passes() throws Throwable {
-        authorizationRule.testCall(this::pass)
+        authorizationRule.given(this::pass)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"))
                 .then(expectNotToFail())
                 .test();
@@ -29,7 +29,7 @@ public class ComponentInheritanceTest extends TestBaseClass {
 
     @Test
     public void passes_users_roles_syntax() throws Throwable {
-        authorizationRule.testCall(this::pass)
+        authorizationRule.given(this::pass)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"), users("foo"))
                 .then(expectNotToFail())
                 .test();
@@ -37,7 +37,7 @@ public class ComponentInheritanceTest extends TestBaseClass {
 
     @Test
     public void fails() throws Throwable {
-        authorizationRule.testCall(this::fail)
+        authorizationRule.given(this::fail)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"))
                 .then(expectException(IllegalStateException.class))
                 .test();