+++ integration-test/src/test/java/fi/vincit/multiusertest/TestUsersCustomRunnerAndException.java
@@ -38,7 +38,7 @@ public class TestUsersCustomRunnerAndException {
 
     @Test
     public void passes() throws Throwable {
-        authorizationRule.testCall(this::pass)
+        authorizationRule.given(this::pass)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"))
                 .then(expectNotToFail())
                 .test();
@@ -46,7 +46,7 @@ public class TestUsersCustomRunnerAndException {
 
     @Test
     public void passes_users_roles_syntax() throws Throwable {
-        authorizationRule.testCall(this::pass)
+        authorizationRule.given(this::pass)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"), users("foo"))
                 .then(expectNotToFail())
                 .test();
@@ -54,7 +54,7 @@ public class TestUsersCustomRunnerAndException {
 
     @Test
     public void fails() throws Throwable {
-        authorizationRule.testCall(this::fail)
+        authorizationRule.given(this::fail)
                 .whenCalledWithAnyOf(roles("ROLE_ADMIN"))
                 .then(expectException(IllegalStateException.class))
                 .test();