+++ junit-5-test/src/test/java/fi/vincit/multiusertest/junit5/JUnit5NewUserTest.java
@@ -50,7 +50,7 @@ public class JUnit5NewUserTest {
 
     @TestTemplate
     public void expectFailureProducer(Authorization authorization) throws Throwable {
-        authorization.testCall(() -> throwIfUserIs(configuredTest.getProducer()))
+        authorization.given(() -> throwIfUserIs(configuredTest.getProducer()))
                 .whenCalledWithAnyOf(UserIdentifiers.producer())
                 .then(expectException(IllegalStateException.class))
                 .test();
@@ -58,7 +58,7 @@ public class JUnit5NewUserTest {
 
     @TestTemplate
     public void expectFailureUserWithProducerRole(Authorization authorization) throws Throwable {
-        authorization.testCall(() -> throwIfUserIs(configuredTest.getConsumer()))
+        authorization.given(() -> throwIfUserIs(configuredTest.getConsumer()))
                 .whenCalledWithAnyOf(UserIdentifiers.withProducerRole())
                 .then(expectException(IllegalStateException.class))
                 .test();
@@ -68,7 +68,7 @@ public class JUnit5NewUserTest {
     public void expectFailureConsumer(Authorization authorization) throws Throwable {
         configuredTest.logInAs(LoginRole.CONSUMER);
 
-        authorization.testCall(() -> throwIfUserRole("role:ROLE_USER"))
+        authorization.given(() -> throwIfUserRole("role:ROLE_USER"))
                 .whenCalledWithAnyOf(roles("ROLE_USER"))
                 .then(expectException(IllegalStateException.class))
                 .test();